infoneige/requirements.txt
Chameleon==2.16
Mako==1.0.0
MarkupSafe==0.23
Paste==1.7.5.1
PasteDeploy==1.5.2
Pygments==1.6
SQLAlchemy==0.9.6
WebOb==1.4
amqp==1.4.5
anyjson==0.3.3
billiard==3.3.0.18
celery==3.1.13
kombu==3.0.21
pyramid==1.5.1
pyramid-chameleon==0.3
pyramid-debugtoolbar==2.1
pyramid-mako==1.0.2
pyramid-tm==0.7
pytz==2014.4
repoze.lru==0.6
stevedore==0.15
suds==0.4
transaction==1.4.3
translationstring==1.1
venusian==1.0
waitress==0.8.9
wsgiref==0.1.2
zope.deprecation==4.1.1
zope.interface==4.1.1
zope.sqlalchemy==0.7.5
git+git://github.com/Vizzuality/cartodb-python.git#egg=cartodb-python
